requirements.txt
# Python 3
# pip install -r requirements.txt

Pillow